input_parser.py
"""
Module for parsing and converting various input formats into pixel grids.
Supports both text and image inputs for creating contribution graph patterns.
"""

from typing import List, Tuple
from PIL import Image

class InputParser:
    """
    Handles the conversion of different input types into pixel grids.
    """
    
    def __init__(self):
        """Initialize the InputParser."""
        pass
    
    def parse_text(self, text: str) -> List[List[int]]:
        """
        Convert text input into a binary pixel grid.
        
        Args:
            text (str): Input text to convert
            
        Returns:
            List[List[int]]: 2D grid of binary values
        """
        pass
    
    def parse_image(self, image_path: str) -> List[List[int]]:
        """
        Convert image input into a binary pixel grid.
        
        Args:
            image_path (str): Path to the input image
            
        Returns:
            List[List[int]]: 2D grid of binary values
        """
        pass 