requirements.txt
python-dotenv==1.0.0
Pillow==10.2.0
GitPython==3.1.42 